Add optional emission of published files
This adds an option to emit published files, along with associated
configuration changes.

The constants file receives a number of new default values for options
related to this feature.

The command builder uses these options to extend the properties object
if the emitPublishedFileds option is true. Note that the PublishUrl
defaults to a uuid.v4.

Finally, the msbuild-runner library is updated to emit the published
files if the emitPublishedFiles flag is true. In order to do this, it
needs access to the top level through2 stream that this plugin creates.
To accomodate this, the call to startMsBuildTask now accepts a stream
parameter, which allows us to pass the through2 stream down and have
files emitted by it. This also requires updates to the unit tests, in
order to accomodate the change of parameter order.