Move the callback to the correct location in code
Since the glob function is asynchronous, we don't want to call the
callback until it has finished executing. With the callback in the
original location, the stream was emitting the end event before the
glob had pushed files into the stream.

This correctly puts the call to the through2 callback inside the glob
callback, and updates the unit tests accordingly.